+++ lib/src/kernels/convolution/impl/mli_krn_group_conv2d_vdsp.h
@@ -205,7 +205,7 @@ MLI_FORCE_INLINE void group_convolution2D(
                     area.clmn_beg = 0;
                     area.clmn_end = out.width - CEIL_DIV(padding_right, stride_width);
                     // compensate for cases where kernel size is larger than input size
-                    area.clmn_end = MAX(CEIL_DIV(padding_left, stride_width), area.clmn_end);
+                    area.clmn_end = MAX(CEIL_DIV(padding_left, stride_width), (int)area.clmn_end);
                 } else {
                     continue;
                 }
